move second fix to MC-259512
