+++ src/client/resources/debugify.client.mixins.json
@@ -44,6 +44,6 @@
   "mixins": [
     "basic.mc176559.ItemStackAccessor",
     "basic.mc206540.EntityMixin",
-    "basic.mc206540.LivingEntityMixin"
+    "basic.mc259512.LivingEntityMixin"
   ]
 }
